backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
chromadb==0.4.15
sentence-transformers==2.2.2
pypdf2==3.0.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
mysql-connector-python==8.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
torch==2.1.0
transformers==4.35.0
datasets==2.14.5
accelerate==0.24.1
pyarrow==12.0.1
openai==0.28.1 